+++ commands/prefix.js
@@ -6,8 +6,11 @@ module.exports = {
 	description: 'Gets the prefixes for this server.',
 	guildOnly: false,
 	args: false,
-	usage: '',
+	usage: '[new prefix]',
 	async execute(args, message, client) {
+		if (args.length) {
+			await client.commands.get('setconf').execute(['prefix', args[0]], message, client);
+		}
 		const prefix = await client.getServerPrefix(message);
 		const msg = `1. ${client.mention}\n2. ${prefix}`;
 		const embed = new YZEmbed('Sebedius Prefixes', msg).setFooter('2 prefixes');