+++ yearzero/YZCombattant.js
@@ -3,21 +3,21 @@ const Util = require('../utils/Util');
 module.exports = class YZCombattant {
 
 	constructor(data) {
+		if(!data.controller) throw new Error('YZCombattant has no controller');
 		/**
 		 * The unique ID of this Combattant.
 		 * @name YZCombattant#id
 		 * @type {string}
 		 * @readonly
 		 */
-		Object.defineProperty(this, 'author', {
-			value: Util.randomID(),
+		Object.defineProperty(this, 'id', {
+			value: Symbol(Util.randomID()),
 			enumerable: true,
 		});
 
 		this.name = data.name || 'Unnamed';
 		this.hidden = data.hidden || false;
 		this.controller = data.controller;
-		this.group = data.group;
 		this.speed = +data.speed || 1;
 		this.hp = +data.hp || 2;
 		this.armor = +data.armor || 0;