+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2020-06-24
+## Added
+- New feature: **Initiative Tracker** with command `!init`.
+
+## Changed
+- Renamed the old `!init` command into `!drawinit`.
+- Resorted all scripts.
+
 ## [2.0.0] - 2020-06-22
 ### Added
 - Support for **TALES FROM THE LOOP** rolls: `!rolltales` or `!rt`.
@@ -36,7 +44,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `!crit` command revamped. See Readme for details.
 - Relooking of the `!help` command.
 - Updated code for `!panic`, `!admin` and the above commands.
-- Renamed `!init` command into `!drawinit`.
 - Future support for different languages.
 
 ### Fixed